+++ octasine/src/parameters/lfo_frequency_ratio.rs
@@ -1,7 +1,7 @@
 use super::utils::*;
 use super::ParameterValue;
 
-const LFO_FREQUENCY_RATIO_STEPS: [f64; 9] = [
+const LFO_FREQUENCY_RATIO_STEPS: [f32; 9] = [
     1.0 / 16.0,
     1.0 / 8.0,
     1.0 / 4.0,
@@ -31,14 +31,11 @@ impl ParameterValue for LfoFrequencyRatioValue {
     fn get(self) -> Self::Value {
         self.0
     }
-    fn new_from_patch(value: f64) -> Self {
-        Self(map_parameter_value_to_step(
-            &LFO_FREQUENCY_RATIO_STEPS,
-            value,
-        ))
+    fn new_from_patch(value: f32) -> Self {
+        Self(map_parameter_value_to_step(&LFO_FREQUENCY_RATIO_STEPS, value) as f64)
     }
-    fn to_patch(self) -> f64 {
-        map_step_to_parameter_value(&LFO_FREQUENCY_RATIO_STEPS, self.0)
+    fn to_patch(self) -> f32 {
+        map_step_to_parameter_value(&LFO_FREQUENCY_RATIO_STEPS, self.0 as f32)
     }
     fn get_formatted(self) -> String {
         format!("{:.04}", self.0)
