+++ octasine/src/parameters/operator_feedback.rs
@@ -2,7 +2,7 @@ use super::{utils::*, ParameterValue};
 use crate::common::OPERATOR_MOD_INDEX_STEPS;
 
 #[derive(Debug, Clone, Copy)]
-pub struct OperatorFeedbackValue(f64);
+pub struct OperatorFeedbackValue(f32);
 
 impl Default for OperatorFeedbackValue {
     fn default() -> Self {
@@ -11,7 +11,7 @@ impl Default for OperatorFeedbackValue {
 }
 
 impl ParameterValue for OperatorFeedbackValue {
-    type Value = f64;
+    type Value = f32;
 
     fn new_from_audio(value: Self::Value) -> Self {
         Self(value)
@@ -19,13 +19,13 @@ impl ParameterValue for OperatorFeedbackValue {
     fn get(self) -> Self::Value {
         self.0
     }
-    fn new_from_patch(value: f64) -> Self {
+    fn new_from_patch(value: f32) -> Self {
         Self(map_parameter_value_to_value_with_steps(
             &OPERATOR_MOD_INDEX_STEPS[..],
             value,
         ))
     }
-    fn to_patch(self) -> f64 {
+    fn to_patch(self) -> f32 {
         map_value_to_parameter_value_with_steps(&OPERATOR_MOD_INDEX_STEPS[..], self.0)
     }
     fn get_formatted(self) -> String {
